Extract ViewportDependentPillarBoxes component
Add a general component to frontend which handles calculation of
pillar boxes dependent of the aspect ratio of a file
