+++ entry_types/scrolled/package/src/contentElements/inlineVideo/InlineVideo.module.css
@@ -1,8 +1,3 @@
-.root {
-  position: relative;
-  max-height: 100vh;
-}
-
 /* Timeline in native video controls is hidden in Storybook to prevent
    Percy from detecting different loading progress as visual
    change. See .storybook/preview-head.html file. */
