+++ entry_types/scrolled/package/src/frontend/VideoPlayer.js
@@ -4,6 +4,7 @@ import {MediaPlayer} from './MediaPlayer';
 
 import classNames from 'classnames';
 import styles from "./VideoPlayer.module.css";
+import {ViewportDependentPillarBoxes} from "./ViewportDependentPillarBoxes";
 
 
 /**
@@ -17,19 +18,13 @@ export function VideoPlayer(props) {
 
   if (videoFile && videoFile.isReady) {
     const processedSources = processSources(videoFile);
-    const css = calculatePadding(videoFile, props.position);
-
     return (
-      <div style={{paddingTop: css.paddingTop}}>
-        <div className={styles.inner}
-             style={{left: css.leftRight, right: css.leftRight}}>
-          <MediaPlayer className={classNames(styles.video_player, {[styles.backdrop]: !props.interactive})}
-                       type={'video'}
-                       sources={processedSources}
-                       {...props}
-          />
-        </div>
-      </div>
+      <ViewportDependentPillarBoxes file={videoFile} position={props.position} children={
+        <MediaPlayer className={classNames(styles.video_player, {[styles.backdrop]: !props.interactive})}
+                     type={'video'}
+                     sources={processedSources}
+                     {...props}/>
+      }/>
     );
   } else {
     return null;
@@ -40,27 +35,6 @@ VideoPlayer.defaultProps = {
   interactive: false,
   controls: true
 };
-
-function calculatePadding(videoFile, position) {
-  const videoAR = (videoFile.height / videoFile.width);
-  let baseCss = {
-    paddingTop: (videoAR * 100) + '%',
-    leftRight: 0
-  }
-
-  if (position === 'full') {
-    const viewportHeight = window.innerHeight;
-    const viewportWidth = window.innerWidth;
-    const viewPortAR = (viewportHeight / viewportWidth);
-    if (viewPortAR < videoAR) {
-      const leftRight = (viewportWidth - (viewportHeight / videoAR)) / 2;
-      baseCss.paddingTop = (viewPortAR * 100) + '%';
-      baseCss.leftRight = leftRight + 'px'
-    }
-  }
-
-  return baseCss;
-}
 
 function processSources(videoFile) {
   return [
