+++ entry_types/scrolled/package/src/frontend/VideoPlayer.module.css
@@ -1,11 +1,3 @@
-.inner {
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-}
-
 .video_player {
   width: 100%;
   height: 100%;
