+++ entry_types/scrolled/package/src/frontend/ViewportDependentPillarBoxes.module.css
@@ -1,0 +1,12 @@
+.root {
+  position: relative;
+  max-height: 100vh;
+}
+
+.inner {
+  position: absolute;
+  top: 0;
+  left: 0; /* will get overwritten inline dependent on viewport */
+  bottom: 0;
+  right: 0; /* will get overwritten inline dependent on viewport */
+}
